Cards Database
Added a script that makes the card parts deck and assigns all the values to each card. I haven't duplicated them yet because I want to wait until everything is finalized to do that so there is only one card of each for now. Images for the cards are just the d20 for now as a placeholder so those should all be replaced before we are ready to have people play.